+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/CommentService.java
@@ -29,5 +29,5 @@ public interface CommentService {
 
     PaginatedCollection<Comment> findCommentsByUser(User user, int pageNumber, int pageSize);
 
-    int getMaxCommentTreeDepth();
+    long getMaxCommentTreeDepth();
 }
